Polishing duplicated code
